blog/01.system/01.wordpress_uploader/requirements.txt
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0
pathlib>=1.0.1